backend/requirements.txt
# ── Web Framework ────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── HTTP Client (async GitHub API calls) ─────────────────────
httpx==0.27.0

# ── Machine Learning ──────────────────────────────────────────
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2

# ── Code Complexity Analysis ──────────────────────────────────
radon==6.0.1

# ── Database ──────────────────────────────────────────────────
supabase==2.4.6

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
